Fix javadoc build on java 11 for main modules.
This requires a HTML 5 compliant <table> in the MersenneTwister javadocs
and a new <caption> element to replace the summary attribute.

Note: The JPMS modules fail. The javadoc tool for java 11 will not support links to older java APIs when building for a Java 9 source due to "packages in the unamed module".