Changed samplers parametric test failure probability.
Only 1 re-run is allowed and the number of samplers is high enough that
spurious failures are realistic for the samplers.

Changed the Continuous samplers failure rate from p=0.0138 to p=0.0016.
This is an effective change of numFailures(n=50, p=0.01) > 2 to
numFailures(n=50, p=0.01) > 3.

Changed the Discrete samplers failure rate from p=0.0000189 to p=0.0016
to match the continuous samplers test. Each test is now assessed at the
p=0.01 level not the p=0.001 level. This is an effective change of
numFailures(n=50, p=0.001) > 2 to numFailures(n=50, p=0.01) > 3.