Update ziggurat test to target regions of the ziggurat.
This targets concave, inflection, and convex overhangs and the tail of
the Gaussian.

This targets concave overhangs, and the tail of the exponential.

Note: This test is still not able to identify minor errors in the sampling algorithm. The test cannot detect the difference between a concave region sampled correctly or sampled using a uniform sample from the lower-left triangle.