Update ziggurat sampler using result of the performance tests
Use a single method to interpolate x and y. This avoids a subtraction in
the interpolation of y.

Avoid recursion in the exponential sampler.

Use shifts to generate unsigned longs. This favours the upper bits of
the uniform deviate.

Each change is either mute or may have a marginal benefit depending on
the platform. Together the sampler should be consistently performant
across JDKs and platforms.

Update tests that manipulate the output from the exponential sampler:
stable sampler test
ziggurat sampler test