Remove badges from examples README.md
The badges do not relate to the examples code.

The build badge is for the main RNG modules. No check is made if the
examples code even builds.

The coverage badge is for the main RNG code. Coverage of the examples is
low to minimal and is not reported on.