RNG-181: LXM family to support SplittableUniformRandomProvider
Create generic RandomStreams class that can stream objects created with
a seed and splittable source of randomness. The seed uses the stream
position mixed with random bits to ensure it is unique within the stream
(up to a size limit of 2^60). Use this feature to support splits in the
LXM family.

Add isSplittable method to RandomSource to use to identify supported
interfaces.

Use of the splittable interface by other modules requires an exception
in RevAPI for exposing an external class in the API. This is similar to
exposure of UniformRandomProvider and is allowed. The change is
non-breaking for binary and source compatibility (see revapi
java.class.externalClassExposedInAPI).