Increase robustness of the LineSampler test
Sporadic failures have been observed for the assertion than the point is
on the line. The test now inverts the point location to a uniform value
in [0, 1) and asserts it matches the output from the RNG.

This removes the requirement for a Chi-square test as the RNG output
should be uniform.